% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_to_irma.R
\name{addCountryName}
\alias{addCountryName}
\title{Add country name to colony locations based on nearest NaturalEarth feature}
\usage{
addCountryName(coldf)
}
\arguments{
\item{coldf}{Data frame with colony locations (must contain columns 'col_lon' and 'col_lat')}
}
\value{
Data frame with added 'sovereignty' and 'country' columns
}
\description{
Add country name to colony locations based on nearest NaturalEarth feature
}
\keyword{internal}
