% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{append_to_nonresponsive}
\alias{append_to_nonresponsive}
\title{Append to nonresponsive list}
\usage{
append_to_nonresponsive(nonresponsive_list, new_nonresponsive, manufacturer)
}
\arguments{
\item{nonresponsive_list}{A list containing tibbles of unresponsive loggers for different manufacturers.
The name of the list element should match the producer name in master_startup (e.g., "Lotek", "MigrateTech").
This can be generated with the \code{load_nonresponsive} function}

\item{new_nonresponsive}{Tibble containing new rows to be appended.}

\item{manufacturer}{Character string of name of manufacturer whose nonresponsive sheet is to be appended to.}
}
\value{
Modified nonresponsive_list
}
\description{
This function appends to the approrpiate sheet in a list of nonresponsive sheets.
It will check for duplicate logger IDs and ensure column ordering matches.
}
\concept{nonresponsive}
