% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_import_utilities.R
\name{check_db_metadata_import}
\alias{check_db_metadata_import}
\title{Check existing rows in database given certain columns from a metadata dataframe}
\usage{
check_db_metadata_import(
  metadata_df,
  table,
  col_names = NULL,
  db_col_names = NULL
)
}
\arguments{
\item{metadata_df}{A dataframe containing metadata to be checked against the database.}

\item{table}{The name of the database table to check against.}

\item{col_names}{A vector of column names from the metadata dataframe to be used for comparison. If NULL, all columns are used.}

\item{db_col_names}{A vector of column names in the database table corresponding to \code{col_names}. If NULL, \code{col_names} are used.}
}
\value{
A logical vector indicating which rows in the metadata dataframe do not exist in the database table.
}
\description{
This function checks for existing rows in a database table based on specified columns from a metadata dataframe.
It constructs a SQL query to compare the values in the dataframe with those in the database table. IF a row from the dataframe
does not exist in the database table based on the specified columns, it is marked as missing.
}
\concept{db_import_utility}
