% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_import_utilities.R
\name{check_deployment_db}
\alias{check_deployment_db}
\title{Check existing deployments in database}
\usage{
check_deployment_db(deployments)
}
\arguments{
\item{deployments}{A tibble containing deployment information from master import metadata.}
}
\value{
A tibble of deployments that are not already present in the database.
}
\description{
Function to check which deployments from the master import sheet are not already present in the database.
It constructs a dataframe of deployment dates and individ IDs, and uses the \code{check_db_metadata_import} function
to determine which deployments are new and need to be added to the database.
}
\concept{db_import_utility}
