% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_import_utilities.R
\name{check_retrieval_db}
\alias{check_retrieval_db}
\title{Check existing retrievals in database}
\usage{
check_retrieval_db(retrievals)
}
\arguments{
\item{retrievals}{A tibble containing retrieval information from master import metadata.}
}
\value{
A tibble of retrievals that are not already present in the database.
}
\description{
Function to check which retrievals from the master import sheet are not already present in the database.
It constructs a dataframe of retrieval dates and individ IDs, and uses the \code{check_db_metadata_import} function
to determine which retrievals are new and need to be added to the database.
}
\concept{db_import_utility}
