% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_import_utilities.R
\name{check_shutdown_db}
\alias{check_shutdown_db}
\title{Check existing shutdowns in database}
\usage{
check_shutdown_db(startup_shutdown)
}
\arguments{
\item{startup_shutdown}{A tibble containing session information from master import startup_shutdown.}
}
\value{
A tibble of shutdowns that are not already present in the database.
}
\description{
Function to check which shutdowns from the master import sheet are not already present in the database.
It constructs a dataframe of session IDs and download dates, and uses the \code{check_db_metadata_import} function
to determine which shutdowns are new and need to be added to the database.
}
\concept{db_import_utility}
