% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_import_utilities.R
\name{check_startups_active_db}
\alias{check_startups_active_db}
\title{Check existing active sessions in database}
\usage{
check_startups_active_db(startup_shutdown, filter_startups = FALSE)
}
\arguments{
\item{startup_shutdown}{A tibble containing session information from master import startup_shutdown.}

\item{filter_startups}{A logical indicating whether to filter out startups associated with active sessions. Default is TRUE.}
}
\value{
If filter_startups is TRUE, a tibble of startups that are associated with active sessions in the database. Otherwise
}
\description{
Function to check which startups from the master import sheet correspond to active sessions in the database.
It constructs a dataframe of session IDs and uses the \code{check_db_metadata_import} function
to determine which startups are associated with active sessions.
}
\concept{db_import_utility}
