% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_import_utilities.R
\name{check_startups_db}
\alias{check_startups_db}
\title{Check existing startups in database}
\usage{
check_startups_db(startup_shutdown)
}
\arguments{
\item{startup_shutdown}{A tibble containing session information from master import startup_shutdown.}
}
\value{
A tibble of startups that are not already present in the database.
}
\description{
Function to check which startups from the master import sheet are not already present in the database.
It constructs a dataframe of session IDs and start times, and uses the \code{check_db_metadata_import} function
to determine which startups are new and need to be added to the database.
}
\concept{db_import_utility}
