% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/startups.R
\name{choose_startup_to_add}
\alias{choose_startup_to_add}
\title{Choose the appropriate startup row to add to the master startup data frame}
\usage{
choose_startup_to_add(
  logger_partner_logger_data,
  all_startups,
  master_import,
  can_dummy_models = NULL
)
}
\arguments{
\item{logger_partner_logger_data}{A data frame containing logger information from the partner metadata, including deployment and retrieval dates.}

\item{all_startups}{A data frame containing all startup rows for the logger, filtered from the startup files.}

\item{master_import}{The loaded master import object.}

\item{can_dummy_models}{Optional data frame of models that can be used to create dummy start times. If NULL, the function will retrieve the list of models from the database.}
}
\value{
A single startup row that is deemed appropriate for adding to the master startup data frame, or NULL if no suitable row is found.
}
\description{
This function selects the most appropriate startup row from the provided startup rows based on the logger's deployment date and other criteria. It handles cases where multiple startup rows exist for a logger and ensures that the selected row is suitable for adding to the master startup data frame.
}
\concept{startups}
