% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correct_models.R
\name{correct_models}
\alias{correct_models}
\title{Correct logger models in metadata based on database values}
\usage{
correct_models(metadata, startups = NULL)
}
\arguments{
\item{metadata}{Metadata dataframe to correct}

\item{startup}{Optional dataframe of logger startups to use for correction if database values are missing}
}
\value{
Corrected metadata dataframe
}
\description{
This function checks the logger models in the metadata against the database values and corrects any mismatches.
It also uses logger startup information to fill in missing logger models if available. Database connection is required
}
\concept{metadata}
