% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_requests.R
\name{create_readme}
\alias{create_readme}
\title{Create SEATRACK documentation README}
\usage{
create_readme(
  request_name,
  file_list,
  species,
  colonies,
  times,
  data_responsible,
  data_dir,
  additional_notes = "",
  additional_files = list(),
  output_file = "README.html"
)
}
\arguments{
\item{request_name}{A string representing the name of the data request.}

\item{file_list}{A character vector of file names included in the data request.}

\item{species}{A character vector of species included in the data request.}

\item{colonies}{A character vector of colonies included in the data request.}

\item{times}{A vector of two dates representing the start and end of the data request}

\item{data_responsible}{Table of people responsible for the data}

\item{data_dir}{A string representing the directory where the data files are located.}

\item{additional_notes}{An optional string containing additional notes to be included in the README file}

\item{additional_files}{An optional list of additional files to include in the README file. Each element of the list should contain the file path to the file to be included and a description.}

\item{output_file}{A string representing the name and path under which the README file will be saved.}
}
\value{
None. The function creates a README file in the specified data directory.
}
\description{
This function generates a README file for a SEATRACK data request package using R Markdown.
The README includes metadata about the data request, such as species, colonies, time span, and additional notes.
}
\concept{data_requests}
