% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_helpers.R
\name{create_test_encounter_data}
\alias{create_test_encounter_data}
\title{Create test encounter data}
\usage{
create_test_encounter_data()
}
\value{
A tibble representing test encounter data.
}
\description{
Generates a test encounter data tibble for use in tests.
}
\keyword{internal}
