% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_helpers.R
\name{create_test_restart_sheet}
\alias{create_test_restart_sheet}
\title{Create test restart sheet
Generates a test restart sheet as a tibble for use in tests.}
\usage{
create_test_restart_sheet()
}
\value{
A tibble representing a test restart sheet.
}
\description{
Create test restart sheet
Generates a test restart sheet as a tibble for use in tests.
}
\keyword{internal}
