% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_helpers.R
\name{create_test_returns_sheet}
\alias{create_test_returns_sheet}
\title{Create test encounter data
Generates a test encounter data tibble for use in tests.}
\usage{
create_test_returns_sheet()
}
\value{
A tibble representing test encounter data.
}
\description{
Create test encounter data
Generates a test encounter data tibble for use in tests.
}
\keyword{internal}
