% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_requests.R
\name{data_request}
\alias{data_request}
\title{Get SEATRACK data request}
\usage{
data_request(
  request_name,
  data_types = c("raw_data", "GLS_positional_data", "IRMA_positional_data",
    "individual_data", "light", "temperature", "activity", "population_maps",
    "logger_info", "immersion"),
  start_year = "2000",
  end_year = format(Sys.Date(), "\%Y"),
  species = NULL,
  colony = NULL,
  age_deployment = "A",
  session_ID = NULL,
  export = TRUE,
  output_dir = NULL,
  additional_notes = "",
  additional_data_files = list(),
  additional_files = list(),
  additional_data = list()
)
}
\arguments{
\item{request_name}{A string representing the name of the data request.}

\item{data_types}{A character vector specifying the types of data to include in the request. Possible values are "GLS_positional_data", "IRMA_positional_data", "individual_data", "light", "temperature", "activity", "population_maps" and "logger_info". Defaults to all types.}

\item{start_year}{An integer representing the start year for the data request.}

\item{end_year}{An integer representing the end year for the data request. Defaults to the current year.}

\item{species}{An optional string specifying the species to filter the data. If NULL, data for all species will be retrieved.}

\item{colony}{An optional string specifying the colony to filter the data. If NULL, data for all colonies will be retrieved.}

\item{age_deployment}{An optional string specifying the age class to filter the data. Possible values are "A" for adults and "C" for juveniles. Defaults to "A".}

\item{session_ID}{Optional list of strings specifying exact session IDs.}

\item{export}{A boolean indicating whether to export the data package as a zip file. If FALSE, the function will return the data as a list instead.}

\item{output_dir}{An optional string specifying the directory where the exported zip file will be saved.}

\item{additional_notes}{An optional string containing additional notes to be included in the README file in the export.}

\item{additional_data_files}{An optional list of additional data files to include in the data directory of the exported zip file. Each element of the list should contain the file path to the file to be included and a description.}

\item{additional_files}{An optional list of additional files to include in the exported zip file. Each element of the list should contain the file path to the file to be included and a description.
If NULL, it will be saved in a default location based on the current year.}

\item{additional_data}{An optional named list of lists containing \code{data}: a data.frame to be included in the export and \code{description}: A string description. Each name corresponds to a data type. This can be used to include data that is not directly available through the SEATRACK database, but is relevant for the data request.}
}
\value{
If \code{export} is \code{TRUE}: None. The function creates a zip file in the specified output directory.
If \code{export} is \code{FALSE}: A named list of data.frames containing the requested data.
}
\description{
This function retrieves SEATRACK data for a specified request, including position data, morphological and breeding data, and recordings of light, temperature, and activity.
If \code{export} is TRUE the \code{export_data_package} function is used to create a zip file containing the data. This includes writing each data type as a compressed parquet file and creating a README file with metadata about the data request.
If \code{export} is FALSE the function returns a named list of data.frames containing the requested data. This can be edited and passed to \code{export_data_package} later if desired.
Requires an active connection to the SEATRACK database.
Currently netCDF files are not included in the data request package, so these have to be injected using \code{additional_data_files} if needed.
}
\examples{
\dontrun{
data_request("Mosbech_120925", 2023, 2025, "Common eider", "Christiansø")
}
}
\concept{data_requests}
