% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{DBImportCollection}
\alias{DBImportCollection}
\title{DBImportCollection Class}
\description{
A convenience class for storing logger sessions alongside the deployments and retrievals that are associated with them.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{sessions}}{SessionBatch containing session information from master import startup_shutdown.}

\item{\code{retrievals}}{Tibble containing retrievals events information from master import metadata}

\item{\code{deployments}}{Tibble containing retrievals events information from master import metadata}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{type}}{The type of import that needs to occur.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-DBImportCollection-new}{\code{DBImportCollection$new()}}
\item \href{#method-DBImportCollection-print}{\code{DBImportCollection$print()}}
\item \href{#method-DBImportCollection-clone}{\code{DBImportCollection$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DBImportCollection-new"></a>}}
\if{latex}{\out{\hypertarget{method-DBImportCollection-new}{}}}
\subsection{Method \code{new()}}{
Create a new DBImportCollection object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DBImportCollection$new(
  sessions = SessionBatch$new(),
  retrievals = tibble(),
  deployments = tibble()
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{sessions}}{A SessionBatch containing session information from master import startup_shutdown.}

\item{\code{retrievals}}{A tibble containing retrievals events information from master import metadata}

\item{\code{deployments}}{A tibble containing deployments events information from master import metadata}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new DBImportCollection object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DBImportCollection-print"></a>}}
\if{latex}{\out{\hypertarget{method-DBImportCollection-print}{}}}
\subsection{Method \code{print()}}{
Print method for DBImportCollection
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DBImportCollection$print()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
The DBImportCollection object invisibly
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DBImportCollection-clone"></a>}}
\if{latex}{\out{\hypertarget{method-DBImportCollection-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DBImportCollection$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
