% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/web_metadata_app.R
\name{deploy_web_metadata_app}
\alias{deploy_web_metadata_app}
\title{Deploy the SEATRACK metadata Shiny app to shinyapps.io.}
\usage{
deploy_web_metadata_app(
  test = TRUE,
  account = NULL,
  token = NULL,
  secret = NULL
)
}
\arguments{
\item{test}{A boolean indicating whether to deploy the app in test mode. Defaults to TRUE, which deploys the app under the name "seatrack_shinyapp_TEST". If FALSE, the app will be deployed under the name "seatrack_shinyapp".}

\item{account}{A string specifying the shinyapps.io account name. If NULL, it will attempt to retrieve the account name from the SHINYAPPS_ACCOUNT environment variable or prompt the user for input.}

\item{token}{A string specifying the shinyapps.io token. If NULL, it will attempt to retrieve the token from the SHINYAPPS_TOKEN environment variable or prompt the user for input.}

\item{secret}{A string specifying the shinyapps.io secret. If NULL, it will attempt to retrieve the secret from the SHINYAPPS_SECRET environment variable or prompt the user for input.}
}
\value{
None. This function deploys the Shiny app to shinyapps.io and does not return a value.
}
\description{
This function deploys the SEATRACK metadata Shiny app to the shinyapps.io platform. It sets up the necessary credentials for deployment, checks for the presence of required data, and then uses the rsconnect package to deploy the app under a specified account and app name.
}
\concept{web_metadata_app}
