% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_export.R
\name{export_appendix_table_excel}
\alias{export_appendix_table_excel}
\title{Export the flextable to an Excel file with formatting}
\usage{
export_appendix_table_excel(
  ft,
  file_path,
  sheet_name = "table",
  body_cell_width = 4.5,
  species_height = 100,
  lme_width = 4
)
}
\arguments{
\item{ft}{A flextable object to be exported}

\item{file_path}{The file path where the Excel file will be saved}

\item{sheet_name}{The name of the worksheet in the Excel file (default is "table")}

\item{body_cell_width}{The width of the body cells (default is 4)}

\item{species_height}{The height of the species header row (default is 100)}

\item{lme_width}{The width of the LME column (default is 4)}
}
\description{
This function exports a flextable object generated by \code{export_appendix_table_excel} to an Excel file, forcing specific formatting such as column widths, row heights, and text rotation.
}
\examples{
\dontrun{
export_appendix_table_excel(ft, "appendix_table.xlsx")
}
}
\concept{annual_report}
