% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gls_workflows.R
\name{export_check_calibration}
\alias{export_check_calibration}
\title{Export calibration check results to Excel}
\usage{
export_check_calibration()
}
\value{
None. The function saves the calibration check results to an Excel file in the specified output directory.
}
\description{
Function to export the results of the calibration check to an Excel file. It uses the \code{gls_check_calibration} function to get the calibration status for all species/colony combinations and then formats the results into a flextable with colored backgrounds based on the proportion of calibrated entries. The resulting table is saved as an Excel file in the specified output directory.
}
\concept{gls_helper}
