% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/field_plan_handling.R
\name{field_plan_combine_locations}
\alias{field_plan_combine_locations}
\title{Combine duplicate locations/species in the field plan sheet}
\usage{
field_plan_combine_locations(field_plan_sheet)
}
\arguments{
\item{field_plan_sheet}{A data frame containing the field plan data}
}
\value{
A data frame with combined locations/species
}
\description{
This function aggregates locations with the same colony name, species and age together by summing all numeric variables.
}
\examples{
\dontrun{
field_plan_combine_locations(field_plan_sheet)
}
}
\concept{field_planning}
