% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_requests.R
\name{find_raw_data}
\alias{find_raw_data}
\title{Find raw data files for a given session ID and data type}
\usage{
find_raw_data(session_ids)
}
\arguments{
\item{session_ids}{A vector of session IDs for which to find raw data files.}
}
\value{
A list containing two elements: \code{local_files}, a data frame of files found in the local directory with their paths, and \code{ftp_files}, a data frame of files found on the FTP server (currently empty, to be implemented).
}
\description{
This function searches for raw data files associated with specific session IDs and a given data type. It first checks a local directory for the files and, if not found, it can be extended to check an FTP server or other storage locations.
}
\concept{data_requests}
