% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/startups.R
\name{get_can_dummy_models}
\alias{get_can_dummy_models}
\title{Get logger models that can have dummy start times}
\usage{
get_can_dummy_models()
}
\value{
A data frame containing logger models that can have dummy start times.
}
\description{
This function retrieves logger models that are eligible for creating dummy start times. It filters out specific producers and logger types that are not suitable for dummy start times.
}
\concept{startups}
