% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/field_plan_handling.R
\name{get_clean_field_plan}
\alias{get_clean_field_plan}
\title{Read and clean the field plan data}
\usage{
get_clean_field_plan(
  field_plan_sheet,
  target_species = NULL,
  use_master_sheets = FALSE,
  all_locations = NULL,
  use_db = TRUE,
  field_year = c(as.numeric(format(Sys.Date(), "\%Y")))
)
}
\arguments{
\item{field_plan_sheet}{A data frame containing the raw field plan data}

\item{target_species}{Vector of species names to use in correcting the different ways partners might write species}

\item{use_master_sheets}{A boolean indicating whether to use master sheets to calculate success.}

\item{all_locations}{An optional data frame containing all location metadata from the master import, if not provided this will be imported.}

\item{use_db}{A boolean indicating whether to use the database to calculate success.}

\item{field_year}{The year to filter the metadata for when using master sheets (default is the current year)}
}
\value{
A cleaned data frame with deployment and retrieval statistics
}
\description{
This function processes the field plan data by selecting relevant columns, renaming them, calculating deployment and retrieval success rates,
and merging with historical data to compute previous deployments. A connection to the SEATRACK database is required.
}
\examples{
\dontrun{
field_plan_sheet <- get_field_plan("path/to/fieldplan.xlsx")
seatrackR::connectSeatrack()
field_plan_clean <- get_clean_field_plan(field_plan_sheet)
}
}
\concept{field_planning}
