% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annual_report_event_tables.R
\name{get_event_table_ft}
\alias{get_event_table_ft}
\title{Get event table as flextable
This function generates a flextable object for an event table based on the field plan and database. It retrieves the relevant data for the specified event type, logger type, target age, and target species, and formats it into a flextable with appropriate styling and formatting for use in the annual report.}
\usage{
get_event_table_ft(
  current_field_plan,
  name,
  target_species,
  event_type = c("deployed", "retrieved"),
  logger_type = c("GLS", "GPS", "GPS-GSM"),
  target_age = c("A", "C")
)
}
\arguments{
\item{current_field_plan}{The current field plan data frame containing information about the planned and actual events.}

\item{name}{The name of the event table to be generated, used for labeling the table.}

\item{target_species}{A vector of target species to include in the table.}

\item{event_type}{The type of event to generate the table for, either "deployed" or "retrieved".}

\item{logger_type}{The type of logger to filter the data by. Either "GLS", "GPS", or "GPS-GSM".}

\item{target_age}{The age class to filter the data by. Either "A" or "C".}
}
\value{
A flextable object representing the event table.
}
\description{
Get event table as flextable
This function generates a flextable object for an event table based on the field plan and database. It retrieves the relevant data for the specified event type, logger type, target age, and target species, and formats it into a flextable with appropriate styling and formatting for use in the annual report.
}
\concept{annual_report}
