% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/field_plan_handling.R
\name{get_history_table}
\alias{get_history_table}
\title{Get historical deployment and retrieval data}
\usage{
get_history_table(
  history_year,
  gps_models,
  gsm_models,
  event_type = c("Deployment", "Retrieval")
)
}
\arguments{
\item{history_year}{The year for which to retrieve historical data (e.g., 2023)}

\item{event_type}{A vector specifying the event type(s) to filter by (defaults to c("Deployment", "Retrieval"))}
}
\value{
A data frame containing historical deployment and retrieval data
}
\description{
This function retrieves historical deployment and retrieval data from the SEATRACK database for a specified year and event type (Deployment or Retrieval).
It filters the data based on the event type, age class, and logger type, and returns a data frame with relevant information.
A connection to the SEATRACK database is required. This should probably be moved to seatrackRdb at some point.
}
\examples{
\dontrun{
seatrackR::connectSeatrack()
history_data <- get_history_table(2023, c("Deployment", "Retrieval"))
}
}
\keyword{internal}
