% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loggers.R
\name{get_logger_from_db}
\alias{get_logger_from_db}
\title{Find logger instances in database}
\usage{
get_logger_from_db(logger_id)
}
\arguments{
\item{logger_id}{logger ID of desired logger}
}
\value{
data frame with logger information
}
\description{
This function tries to find a logger ID in the database. It returns a data frame with the logger information.
}
\concept{loggers}
