% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_master_import_path}
\alias{get_master_import_path}
\title{Get the path of the master import file}
\usage{
get_master_import_path(colony, use_stored = TRUE)
}
\arguments{
\item{colony}{A character string specifying the name of the colony.}

\item{use_stored}{If TRUE, use a pre-existing path, rather than searching for a new one. Defaults to TRUE.}
}
\value{
A character string representing the path to the master import file.
}
\description{
This function constructs a path to the master import file for a given colony.
On finding the path to a colony's master import sheets, it will be stored in the internal environment for later use.
}
\examples{
\dontrun{
get_master_import_folder("ColonyA")
}
}
\concept{metadata}
