% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{get_master_import_paths}
\alias{get_master_import_paths}
\title{Get master import paths}
\usage{
get_master_import_paths()
}
\value{
A named list of paths, where the names are colony names and the values are file paths to the master import sheets for those colonies.
}
\description{
This function retrieves the currently stored master import paths.
}
\concept{metadata}
