% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_import_prep.R
\name{get_open_session_dates}
\alias{get_open_session_dates}
\title{Get open session dates
Convenience function to get the open date for a set of logger sessions.
If the shutdown date is available, use that. If not, use the download date. If both exist, use the earlier date.
If there is another session from the same logger, use the start of that.
If neither are available, use a date far in the future to indicate the session is still open.}
\usage{
get_open_session_dates(sessions)
}
\arguments{
\item{sessions}{A tibble containing session information from master import startup_shutdown.}
}
\value{
A vector of POSIXct dates representing the open dates for the sessions.
}
\description{
Get open session dates
Convenience function to get the open date for a set of logger sessions.
If the shutdown date is available, use that. If not, use the download date. If both exist, use the earlier date.
If there is another session from the same logger, use the start of that.
If neither are available, use a date far in the future to indicate the session is still open.
}
\concept{utility}
