% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gls_workflows.R
\name{gls_calibrate_all}
\alias{gls_calibrate_all}
\title{Calibrate all species/colony combinations found in seatrack logger data import folder}
\usage{
gls_calibrate_all(
  no_pos_only = TRUE,
  rerun_existing = FALSE,
  include_existing = TRUE,
  filter_plots = FALSE,
  rerun_existing_plots = FALSE,
  n_workers = 4
)
}
\arguments{
\item{no_pos_only}{Logical indicating whether to include only loggers without position data in the database. Default is TRUE.}

\item{rerun_existing}{Logical indicating whether to rerun calibration for loggers that already have calibration data. Default is FALSE.}

\item{include_existing}{Logical indicating whether to include existing calibration data in the final output. Default is TRUE.}

\item{filter_plots}{Logical indicating whether to export filter plots. Default is FALSE.}

\item{rerun_existing_plots}{Logical indicating whether to rerun calibration for loggers that already have calibration plots. Default is FALSE.}

\item{n_workers}{Integer specifying the number of worker processes to use for parallel processing. Default is 4.}
}
\value{
None. The function saves the prepared calibration data to the specified output directory.
}
\description{
Uses hard coded file paths to call gls_calibrate_species_colony for all species/colony combinations found in seatrack database for loggers in import folder.
}
\concept{gls_helper}
