% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gls_workflows.R
\name{gls_calibrate_file}
\alias{gls_calibrate_file}
\title{Prepare GLS files for calibration}
\usage{
gls_calibrate_file(
  id_year_model,
  no_pos_only = TRUE,
  rerun_existing = TRUE,
  include_existing = TRUE,
  filter_plots = TRUE,
  rerun_existing_plots = TRUE
)
}
\arguments{
\item{id_year_model}{Character string in the format "loggerID_year_loggermodel" to filter for a specific logger, year and model.}

\item{no_pos_only}{Logical indicating whether to include only loggers without position data in the database. Default is TRUE.}

\item{rerun_existing}{Logical indicating whether to rerun calibration for loggers that already have calibration data. Default is TRUE.}

\item{include_existing}{Logical indicating whether to include existing calibration data in the final output. Default is TRUE.}

\item{filter_plots}{Logical indicating whether to export filter plots. Default is FALSE.}

\item{rerun_existing_plots}{Logical indicating whether to rerun calibration for loggers that already have calibration plots. Default is TRUE.}
}
\value{
None. The function saves the prepared calibration data to the specified output directory.
}
\description{
Uses hard coded file paths to call gls_prepare_calibration. Filters metadata passed to gls_prepare_calibration by file names.
}
\concept{gls_helper}
