% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gls_workflows.R
\name{gls_check_calibration}
\alias{gls_check_calibration}
\title{Check calibration status for all species/colony combinations}
\usage{
gls_check_calibration()
}
\value{
A dataframe summarizing the calibration status for each species/colony combination, including the number of calibrated, uncalibrated, and valid entries, as well as the total number of entries and the path to the calibration file.
}
\description{
Function to check the calibration status for all species/colony combinations in the Sea Track database. It looks for existing calibration files and summarizes the number of calibrated, uncalibrated, and valid entries for each combination.
}
\concept{gls_helper}
