% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gls_upload.R
\name{gls_import}
\alias{gls_import}
\title{Import GLS position data to the SEATRACK database}
\usage{
gls_import(
  species = NULL,
  colony = NULL,
  chunk_size = 10,
  gls_directory_path = file.path(the$sea_track_folder,
    "Database\\\\Imports_Logger data\\\\Output_GLSpositions")
)
}
\arguments{
\item{species}{A character vector of species names to filter the import. If NULL, all species directories in the specified GLS directory will be processed. Default is NULL.}

\item{colony}{A character vector of colony names to filter the import. If NULL, all colony directories within the specified species directories will be processed. Default is NULL.}

\item{chunk_size}{Integer specifying the number of sessions to upload in each chunk. Default is 10.}

\item{gls_directory_path}{Path to the directory containing the prepared GLS position data files. Default is \verb{Database\\\\Imports_Logger data\\\\Output_GLSpositions} within the SEATRACK folder.}
}
\value{
None. The function uploads the position data to the SEATRACK database.
}
\description{
Function to import prepared GLS position data from a specified directory to the SEATRACK database.
}
\concept{gls_helper}
