% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gls_workflows.R
\name{gls_prepare_calibration}
\alias{gls_prepare_calibration}
\title{Prepare GLS calibration data using seatrack database}
\usage{
gls_prepare_calibration(
  import_directory,
  output_directory,
  species = NULL,
  colony = NULL,
  no_pos_only = TRUE,
  existing_calibration_dir = NULL,
  rerun_existing = TRUE,
  include_existing = TRUE,
  filter_plots = FALSE,
  rerun_existing_plots = FALSE
)
}
\arguments{
\item{import_directory}{Path to the directory containing GLS files.}

\item{output_directory}{Path to the directory where the prepared calibration data will be saved.}

\item{species}{Species name to filter metadata. Default is NULL (no filtering).}

\item{colony}{Colony name to filter metadata. Default is NULL (no filtering).}

\item{no_pos_only}{Logical indicating whether to include only loggers without position data in the database. Default is TRUE.}

\item{existing_calibration_dir}{Directory containing existing calibration data. Default is NULL. If provided, calibration data from this directory will be merged.}

\item{rerun_existing}{Logical indicating whether to rerun calibration for loggers that already have calibration data. Default is TRUE.}

\item{include_existing}{Logical indicating whether to include existing calibration data in the final output. Default is TRUE.}

\item{filter_plots}{Logical indicating whether to export filter plots. Default is FALSE.}
}
\value{
None. The function saves the prepared calibration data to the specified output directory.
}
\description{
Function to prepare GLS calibration data for use with seatrackRgls, based on GLS files in the import directory and metadata from the Sea Track database.
}
\concept{gls_helper}
