% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gls_workflows.R
\name{gls_prepare_file_upload}
\alias{gls_prepare_file_upload}
\title{Prepare a single GLS position data file for database upload}
\usage{
gls_prepare_file_upload(gls_file_path)
}
\arguments{
\item{gls_file_path}{Path to the GLS position data file.}
}
\value{
A dataframe containing the prepared GLS position data ready for database upload.
}
\description{
Function to prepare a single GLS position data file for upload to the Sea Track database.
}
\concept{gls_helper}
