% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gls_workflows.R
\name{gls_prepare_folder_upload}
\alias{gls_prepare_folder_upload}
\title{Prepare GLS position data for database upload}
\usage{
gls_prepare_folder_upload(gls_directory_path)
}
\arguments{
\item{gls_directory_path}{Path to the directory containing GLS position data files.}
}
\value{
A dataframe containing the prepared GLS position data ready for database upload.
}
\description{
Function to prepare GLS position data from a specified directory for upload to the Sea Track database.
}
\concept{gls_helper}
