% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gls_workflows.R
\name{gls_process_positions}
\alias{gls_process_positions}
\title{Process GLS position data using seatrackRgls}
\usage{
gls_process_positions(
  no_pos_only = TRUE,
  filter_plots = TRUE,
  skip_existing_files = TRUE,
  n_workers = 4
)
}
\arguments{
\item{no_pos_only}{Logical indicating whether to include only loggers without position data in the database. Default is TRUE.}

\item{filter_plots}{Logical indicating whether to export filter plots. Default is FALSE.}

\item{skip_existing_files}{Logical indicating whether to skip processing for files that already have processed position data in the output directory. Default is TRUE.}

\item{n_workers}{Integer specifying the number of worker processes to use for parallel processing.}
}
\value{
None. The function saves the processed position data to \verb{Database\\\\Imports_Logger data\\\\Output_GLSpositions}
}
\description{
Function to process calibrated GLS position data using seatrackRgls.
}
\concept{gls_helper}
