% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gls_functions.R
\name{gls_seatrack_calibration}
\alias{gls_seatrack_calibration}
\title{Handle existing GLS calibration data}
\usage{
gls_seatrack_calibration(metadata_path, split_years = "06-01")
}
\arguments{
\item{metadata_path}{Path to metadata Excel file}

\item{split_years}{Character string indicating the month and day to split years for calibration (e.g., "06-01" for June 1st). Default is "06-01".}
}
\value{
A list with two dataframes: calibration_data and extra_metadata
}
\description{
Function to reshape existing seatrack GLS calibration settings for immediate use in seatrackRgls
}
\concept{gls_helper}
