% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gls_upload.R
\name{gls_upload_pos_data}
\alias{gls_upload_pos_data}
\title{Upload GLS position data to the SEATRACK database}
\usage{
gls_upload_pos_data(pos_list, chunk_size = 10)
}
\arguments{
\item{pos_list}{A list of dataframes containing the prepared GLS position data, where each dataframe corresponds to a session and is named with the session ID.}

\item{chunk_size}{Integer specifying the number of sessions to upload in each chunk. Default is 10.}
}
\value{
None. The function uploads the position data to the SEATRACK database.
}
\description{
Function to upload prepared GLS position data to the Sea Track database.
}
\concept{gls_helper}
