% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activity_immersion_data_functions.R
\name{handle_immersion_lotek}
\alias{handle_immersion_lotek}
\title{Handle immersion data from Lotek logger}
\usage{
handle_immersion_lotek(filepath)
}
\arguments{
\item{filepath}{The full path to the Lotek logger file to be processed.}
}
\value{
A cleaned data frame with date-time, conductivity, and standardized conductivity information, or NULL if the file fails quality checks.
}
\description{
This function processes immersion data from a Lotek logger file. It reads the file, extracts date-time and conductivity information, performs error checking for data quality, and returns a cleaned data frame with standardized conductivity values.
}
\concept{activity_db_prep}
