% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activity_light_data_functions.R
\name{handle_light_lotek}
\alias{handle_light_lotek}
\title{Handle light data from Lotek logger}
\usage{
handle_light_lotek(filepath)
}
\arguments{
\item{filepath}{The full path to the Lotek logger file to be processed.}
}
\value{
A cleaned data frame with date-time, light, and standardized light information, or NULL if the file fails quality checks.
}
\description{
This function processes light data from a Lotek logger file. It reads the file, extracts date-time and light information, performs error checking for data quality, and returns a cleaned data frame with standardized light values.
}
\concept{activity_db_prep}
