% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activity_temperature_functions.R
\name{handle_temperature_lotek}
\alias{handle_temperature_lotek}
\title{Handle temperature data from Lotek logger}
\usage{
handle_temperature_lotek(filepath)
}
\arguments{
\item{filepath}{The full path to the Lotek logger file to be processed.}
}
\value{
A cleaned data frame with date-time, temperature, and standardized temperature information, or NULL if the file fails quality checks.
}
\description{
This function processes temperature data from a Lotek logger file. It reads the file, extracts date-time and temperature information, performs error checking for data quality, and returns a cleaned data frame with standardized temperature values.
}
\concept{activity_db_prep}
