% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activity_immersion_data_functions.R
\name{load_immersion_data}
\alias{load_immersion_data}
\title{Load immersion data from file}
\usage{
load_immersion_data(file_info)
}
\arguments{
\item{file_info}{A dataframe containing information about the file to be processed, including session_id, filename, individ_id, deployment_date, retrieval_date, full_path, and extension.}
}
\value{
A cleaned data frame with date-time, conductivity, and standardized conductivity information, or NULL if the file fails quality checks.
}
\description{
This function loads immersion data from a specified file, processes it according to the file type (Lotek or Migrate), and returns a cleaned data frame with date-time and conductivity information. The function includes error checking for data quality and consistency.
}
\concept{activity_db_prep}
