% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{load_nonresponsive_sheet}
\alias{load_nonresponsive_sheet}
\title{Load nonresponsive logger sheet for current year}
\usage{
load_nonresponsive_sheet(file_path)
}
\arguments{
\item{file_path}{String indicating from where the file should be loaded from.}
}
\value{
A tibble containing the unresponsive logger data.
}
\description{
This function loads the record of unresponsive loggers. If the filepath provided does not exist, it initialises new sheets.
}
\concept{nonresponsive}
