% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{LoadedWB}
\alias{LoadedWB}
\title{LoadedMetadata Class}
\description{
A convenience class for handling metadata loaded from workbooks.
}
\seealso{
Other classes: 
\code{\link{DBImportCollection}},
\code{\link{LoadedWBCollection}},
\code{\link{SessionBatch}}
}
\concept{classes}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{data}}{A list of tibbles}

\item{\code{wb}}{A workbook object}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{path}}{The path of the workbook}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-LoadedWB-new}{\code{LoadedWB$new()}}
\item \href{#method-LoadedWB-print}{\code{LoadedWB$print()}}
\item \href{#method-LoadedWB-clone}{\code{LoadedWB$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LoadedWB-new"></a>}}
\if{latex}{\out{\hypertarget{method-LoadedWB-new}{}}}
\subsection{Method \code{new()}}{
Create a new LoadedWB object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LoadedWB$new(data = list(), wb = openxlsx2::wb_workbook())}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{A list of tibbles}

\item{\code{wb}}{A workbook object}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new LoadedWB object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LoadedWB-print"></a>}}
\if{latex}{\out{\hypertarget{method-LoadedWB-print}{}}}
\subsection{Method \code{print()}}{
Print method for LoadedWB
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LoadedWB$print()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
The LoadedWB object invisibly
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LoadedWB-clone"></a>}}
\if{latex}{\out{\hypertarget{method-LoadedWB-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LoadedWB$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
