% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{LoadedWBCollection}
\alias{LoadedWBCollection}
\title{LoadedWBCollection Class}
\description{
A convenience class for handling collections of LoadedWB objects.
}
\seealso{
Other classes: 
\code{\link{DBImportCollection}},
\code{\link{LoadedWB}},
\code{\link{SessionBatch}}
}
\concept{classes}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{sheets_list}}{A list of LoadedWB objects}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{all_paths}}{The paths of all workbooks in the collection}

\item{\code{all_names}}{The names of all workbooks in the collection}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-LoadedWBCollection-new}{\code{LoadedWBCollection$new()}}
\item \href{#method-LoadedWBCollection-names}{\code{LoadedWBCollection$names()}}
\item \href{#method-LoadedWBCollection-print}{\code{LoadedWBCollection$print()}}
\item \href{#method-LoadedWBCollection-clone}{\code{LoadedWBCollection$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LoadedWBCollection-new"></a>}}
\if{latex}{\out{\hypertarget{method-LoadedWBCollection-new}{}}}
\subsection{Method \code{new()}}{
Create a new LoadedWBCollection object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LoadedWBCollection$new(sheets_list = list())}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{sheets_list}}{A list of LoadedWB objects}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new LoadedWBCollection object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LoadedWBCollection-names"></a>}}
\if{latex}{\out{\hypertarget{method-LoadedWBCollection-names}{}}}
\subsection{Method \code{names()}}{
Get the names of the sheets in the collection
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LoadedWBCollection$names()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A character vector of sheet names
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LoadedWBCollection-print"></a>}}
\if{latex}{\out{\hypertarget{method-LoadedWBCollection-print}{}}}
\subsection{Method \code{print()}}{
Print method for LoadedWBCollection
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LoadedWBCollection$print()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
The LoadedWBCollection object invisibly
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LoadedWBCollection-clone"></a>}}
\if{latex}{\out{\hypertarget{method-LoadedWBCollection-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LoadedWBCollection$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
