% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{modify_logger_status}
\alias{modify_logger_status}
\title{Modify logger status}
\usage{
modify_logger_status(
  logger_id,
  new_data = list(),
  master_sheet = NULL,
  all_master_sheet = NULL
)
}
\arguments{
\item{logger_id}{logger serial number}

\item{new_data}{named list of new data to be inserted, where the name of each element corresponds to a column in the sheet.}

\item{master_sheet}{master sheet in which the logger will be found. If not provided, all sheets will be checked.}

\item{all_master_sheet}{If master_sheet is not provided, all sheets will be checked. To save these being loaded every time, they can be provided using \code{load_all_master_import(combine = FALSE)}}
}
\value{
modified version of master_sheet
}
\description{
Function to modify a logger status, such as download_date and shutdown_date if the logger session is open.
}
\concept{loggers}
