% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{nonresponsive_from_master}
\alias{nonresponsive_from_master}
\title{Get all nonresponsive loggers and export}
\usage{
nonresponsive_from_master(
  all_metadata_combined,
  nonresponsive_list,
  start_date = NULL,
  end_date = Inf
)
}
\arguments{
\item{all_metadata_combined}{Tibble containing combined startup information from master startup sheets, as generated by \code{load_all_master_import(TRUE)}}

\item{nonresponsive_list}{A list containing tibbles of unresponsive loggers for different manufacturers.
The name of the list element should match the producer name in master_startup (e.g., "Lotek", "MigrateTech").
This can be generated with the \code{load_nonresponsive} function}

\item{start_date}{Date or character string specifying the start date for considering nonresponsive loggers. If NULL, defaults to January 1st of the current year.}

\item{end_date}{Date or character string specifying the end date for considering nonresponsive loggers. If NULL, defaults to Inf (no end date).}
}
\value{
Updated list of nonresponsive logger sheets
}
\description{
Checks the master startup sheets for loggers marked as nonresponsive handled between certain dates
and returns per-manufacturer nonresponsive sheets.
}
\concept{nonresponsive}
