% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activity_utilities.R
\name{parse_migratetech_header}
\alias{parse_migratetech_header}
\title{Parse the header of a migratetech file}
\usage{
parse_migratetech_header(filepath)
}
\arguments{
\item{filepath}{The full path to the migratetech file to be processed.}
}
\value{
A list containing the vector of column names and the wet number.
}
\description{
Convert the header of a migratetech file into a vector of column names and extract the wet number.
}
\concept{activity_db_prep}
