% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ColonyRepr_supporting_functions.R
\name{prep_SEATRACK_ColonyRepr}
\alias{prep_SEATRACK_ColonyRepr}
\title{Extracts data from the SEATRACK DB and prepares it for the ColonyRepr functions}
\usage{
prep_SEATRACK_ColonyRepr(species, colony)
}
\arguments{
\item{species}{specify the species you'd like to extract the data from}

\item{colony}{specify the colony you'd like to select the data from}
}
\value{
position data from the specified species-colony, ready to be inserted into ColonyRepr functions
}
\description{
This function will extract gls positional data for the specified species-colony. An active connection with the data base is required for this.
It will then prepare the extracted data to be used in the ColonyRepr functions
}
\concept{ColonyRepr}
