% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ColonyRepr_supporting_functions.R
\name{Prep_SEATRACK_data_for_ColonyRepr}
\alias{Prep_SEATRACK_data_for_ColonyRepr}
\title{Extracts data from the SEATRACK DB and prepares it for the ColonyRepr functions}
\usage{
Prep_SEATRACK_data_for_ColonyRepr(species, colony)
}
\arguments{
\item{species}{specify the species you'd like to extract the data from}

\item{colony}{specify the colony you'd like to select the data from}
}
\value{
position data from the specified species-colony, ready to be inserted into ColonyRepr functions
}
\description{
Extracts data from the SEATRACK DB and prepares it for the ColonyRepr functions
}
\concept{ColonyRepr}
