% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_import_prep.R
\name{prepare_master_sheet_for_db}
\alias{prepare_master_sheet_for_db}
\title{Prepare master sheet for database import}
\usage{
prepare_master_sheet_for_db(master_sheets)
}
\arguments{
\item{master_sheets}{A LoadedWBCollection object containing the master import sheets.}
}
\value{
A vector of DBImportCollection objects representing the prepared session batches for database import.
}
\description{
This function prepares the master import sheet for database import by processing the metadata and startup/shutdown information.
It checks the database for existing sessions and categorizes them into sessions to be closed, opened, or both opened and closed.
It then prepares session batches for database import. Metadata is linked to sessions within these batches.
}
\concept{db_import}
