% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_import_prep.R
\name{prepare_session_batch}
\alias{prepare_session_batch}
\title{Prepare session batch for database import}
\usage{
prepare_session_batch(session_batch, metadata)
}
\arguments{
\item{session_batch}{A SessionBatch object to be prepared for database import.}

\item{metadata}{A dataframe containing metadata information for the sessions.}
}
\value{
DBImportCollection object representing the prepared session batches for database import.
}
\description{
This function prepares a session batch for database import by processing each batch and extracting relevant metadata.
It adds deployments and retrievals associated with the sessions in each batch.
In cases where sessions have multiple deployments or retrievals, those sessions are removed from the batch and a warning is logged.
}
\concept{db_import}
