% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_import_prep.R
\name{prepare_session_batches}
\alias{prepare_session_batches}
\title{Prepare session batches for database import}
\usage{
prepare_session_batches(session_batches, metadata)
}
\arguments{
\item{session_batches}{A vector of SessionBatch objects to be prepared for database import.}

\item{metadata}{A dataframe containing metadata information for the sessions.}
}
\value{
A vector of DBImportCollection objects representing the prepared session batches for database import.
}
\description{
This function prepares a list of session batches for database import by processing each batch and extracting relevant metadata.
It filters out any batches that do not contain sessions and returns a list of prepared session batches.
It adds deployments and retrievals associated with the sessions in each batch.
In cases where sessions have multiple deployments or retrievals, those sessions are removed from the batch and a warning is logged.
}
\concept{db_import}
