% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_import_prep.R
\name{prepare_session_retrievals}
\alias{prepare_session_retrievals}
\title{Get session retrievals}
\usage{
prepare_session_retrievals(
  sessions,
  metadata,
  filter_sessions = FALSE,
  report_missing = FALSE,
  duplicate_warnings = TRUE
)
}
\arguments{
\item{sessions}{A tibble containing session information from master import startup_shutdown.}

\item{metadata}{A dataframe containing metadata information for the sessions.}

\item{filter_sessions}{Logical indicating whether to filter out sessions with no retrievals.}

\item{report_missing}{Logical indicating whether to report retrievals that are not associated with any sessions.}

\item{duplicate_warnings}{Logical indicating whether to report duplicate retrievals}
}
\value{
A list containing a dataframe of retrievals and a tibble of sessions.
}
\description{
Function to get retrievals associated with a set of logger sessions. The function will look for retrievals that occurred between the session start time and the session end time (download/shutdown).
If the session is open (no download/shutdown date), the function will use a date far in the future to indicate the session is still open.
If there are ambiguities (multiple retrievals for a single session), those sessions will be removed and a warning will be logged.
}
\concept{db_import_prep}
