% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_import.R
\name{push_master_import_file_to_db}
\alias{push_master_import_file_to_db}
\title{Push master import file to database}
\usage{
push_master_import_file_to_db(colony = NULL, file_path = NULL)
}
\arguments{
\item{colony}{Optional colony name to filter the master import file.}

\item{file_path}{Optional file path to a master import Excel file.}
}
\description{
Function to push a master import file to the database. The function loads the master import file, prepares the data for database import, and then pushes the data to the database using the \code{push_db_import_collections} function.
}
\concept{db_import}
