% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_import.R
\name{push_shutdowns}
\alias{push_shutdowns}
\title{Push shutdowns to database}
\usage{
push_shutdowns(shutdown, dry_run = FALSE)
}
\arguments{
\item{shutdown}{A tibble containing session information from a DBImportCollection object.}

\item{dry_run}{Logical indicating whether to perform a dry run (default is FALSE). If TRUE, no data will be written to the database.}
}
\description{
Function to push shutdowns from a master import sheet to the database. It checks the database for existing shutdowns and only imports new shutdowns to avoid duplicates.
It also checks that the loggers being shutdown have an associated retrieval in the database if they have a deployment.
It also checks that loggers without deployments are not being shutdown if they are marked as 'Successfully downloaded' or 'Nonresponsive'.
}
\concept{db_import}
