% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_import_utilities.R
\name{R_df_to_db_values}
\alias{R_df_to_db_values}
\title{Convert a dataframe of R values to a string of database values for SQL insertion}
\usage{
R_df_to_db_values(db_cols, db_types = rep("", length(db_cols)))
}
\arguments{
\item{db_cols}{A dataframe containing the columns to be converted.}

\item{db_types}{A named vector specifying the database types for each column. Default is an empty string for each column.}
}
\value{
A string of database values formatted for SQL insertion.
}
\description{
This function takes a dataframe of R values and converts each value to its corresponding database representation.
It then constructs a string of database values formatted for SQL insertion.
}
\concept{db_import_utility}
