% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{run_app}
\alias{run_app}
\title{Run the Shiny App}
\usage{
run_app(
  settings_path = file.path(getwd(), "seatrackRtools_app"),
  log_path = file.path(getwd(), "seatrackRtools_app", "logs"),
  test = FALSE
)
}
\arguments{
\item{settings_path}{A string specifying the path to the settings directory for the app. Defaults to a "seatrackRtools_app" directory in the current working directory.}

\item{log_path}{A string specifying the path to the logs directory for the app. Defaults to a "logs" directory within the "seatrackRtools_app" directory in the current working directory.}

\item{test}{A boolean indicating whether to run the app in test mode. Defaults to FALSE. If TRUE, the app will use the test database.}
}
\value{
None. This function launches the Shiny app and does not return a value.
}
\description{
This function launches the Shiny application for seatrackRtools. It sets up the necessary options for the app, including paths for settings and logs, and then runs the app from the specified directory within the package.
}
\concept{shiny_app}
