% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ColonyRepr_supporting_functions.R
\name{run_ColonyRepr_app}
\alias{run_ColonyRepr_app}
\title{Run the ColonyRepr shiny app}
\usage{
run_ColonyRepr_app(SEATRACK_path = the$sea_track_folder)
}
\description{
Run the ColonyRepr shiny app
}
\concept{ColonyRepr}
