% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/web_metadata_app.R
\name{run_web_metadata_app}
\alias{run_web_metadata_app}
\title{Run the Shiny app for visualising SEATRACK metadata.}
\usage{
run_web_metadata_app()
}
\value{
None. This function launches the Shiny app and does not return a value
}
\description{
This function launches the SEATRACK web metadata app locally.
}
\concept{web_metadata_app}
