% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{save_master_sheet}
\alias{save_master_sheet}
\title{Save a master sheet to an Excel file.}
\usage{
save_master_sheet(new_master_sheets, filepath = NULL, modified_only = FALSE)
}
\arguments{
\item{new_master_sheets}{A LoadedWB object.}

\item{filepath}{A string specifying the path and name of the Excel file to be created. If NULL will be path of the loaded sheet.}

\item{modified_only}{Only save the file if the sheet has been flagged as modified.}
}
\value{
The function returns the updated LoadedWB object after saving.
}
\description{
This function writes the provided data frame (\code{new_master_sheets}) to an Excel file
specified by \code{filename}.
}
\examples{
\dontrun{
save_master_sheet(new_master_sheets, "output.xlsx")
}

}
\concept{metadata}
