% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seatrackRtools-package.R
\docType{package}
\name{seatrackRtools-package}
\alias{seatrackRtools}
\alias{seatrackRtools-package}
\title{seatrackRtools: Internal utility Functions for Seatrack}
\description{
Internal utility functions to assist in the handling of data as part of the seatrack project.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://ninanor.github.io/seatrackRtools}
}

}
\author{
\strong{Maintainer}: Julian Evans \email{julian.evans@nina.no}

Other contributors:
\itemize{
  \item Lars Ursem \email{lars.ursem@npolar.no} [contributor]
}

}
\keyword{internal}
