% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{SessionBatch}
\alias{SessionBatch}
\title{SessionBatch Class}
\description{
A convenience class for storing a set of logger sessions along with the type of database import they require.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{sessions}}{Tibble containing session information from master import startup_shutdown.}

\item{\code{type}}{The type of import that needs to occur. Must be either "close_only", "open_only" or "open_and_close".}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-SessionBatch-new}{\code{SessionBatch$new()}}
\item \href{#method-SessionBatch-print}{\code{SessionBatch$print()}}
\item \href{#method-SessionBatch-clone}{\code{SessionBatch$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SessionBatch-new"></a>}}
\if{latex}{\out{\hypertarget{method-SessionBatch-new}{}}}
\subsection{Method \code{new()}}{
Create a new SessionBatch object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SessionBatch$new(
  sessions = tibble(),
  type = c("close_only", "open_only", "open_and_close")
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{sessions}}{A tibble containing session information from master import startup_shutdown.}

\item{\code{type}}{The type of import that needs to occur. Must be either "close_only", "open_only" or "open_and_close".}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new SessionBatch object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SessionBatch-print"></a>}}
\if{latex}{\out{\hypertarget{method-SessionBatch-print}{}}}
\subsection{Method \code{print()}}{
Print method for SessionBatch
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SessionBatch$print()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
The SessionBatch object invisibly
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SessionBatch-clone"></a>}}
\if{latex}{\out{\hypertarget{method-SessionBatch-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SessionBatch$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
