% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{set_master_import_paths}
\alias{set_master_import_paths}
\title{Set master import paths}
\usage{
set_master_import_paths(location_paths = NULL)
}
\arguments{
\item{location_paths}{A named list of paths, where the names are colony names and the values are file paths to the master import sheets for those colonies.}
}
\description{
This function allows setting the master import paths directly, for example if they have been discovered through other means. This will overwrite any previously stored paths.
}
\concept{metadata}
