% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nonresponsive.R
\name{set_reconstructed}
\alias{set_reconstructed}
\title{Set reconstructed download type for nonresponsive loggers}
\usage{
set_reconstructed(all_master_sheets = NULL)
}
\arguments{
\item{all_master_sheets}{Optional list of master startup sheets. If NULL, the function will load all master startup sheets using \code{load_all_master_import()}.}
}
\value{
No return value.
}
\description{
This function checks the master startup sheets for loggers marked as nonresponsive and updates their download type to "Reconstructed" in both the master startup sheets and the database.
}
\concept{nonresponsive}
