% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.R
\name{set_sea_track_folder}
\alias{set_sea_track_folder}
\title{Set the base directory for the sea track folder}
\usage{
set_sea_track_folder(
  dir,
  language = "English_United Kingdom",
  save_path = TRUE
)
}
\arguments{
\item{dir}{A character string specifying the path to the base directory.}

\item{language}{Character string specifying system language to add utf8 encoding to.}

\item{save_path}{Save seatrack folder to renviron to allow reuse.}
}
\value{
None
}
\description{
This function sets a global variable used by other functions.
It also sets system locale to allow the handling of Norwegian characters in filenames.
}
\examples{
\dontrun{
set_sea_track_folder("/path/to/sea/track/folder")
}
}
\concept{setup}
