% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{setup_app_logs}
\alias{setup_app_logs}
\title{Set up logs for shiny app
This function initializes logging for the Shiny app, creating separate log files for general logs, errors, and successes. It ensures that the specified log directory exists and starts logging with appropriate settings for each log type.}
\usage{
setup_app_logs(
  silent = FALSE,
  log_path = NULL,
  log_names = list(global = paste0("seatrackRtools_log_", format(Sys.time(),
    "\%Y-\%m-\%d_\%H\%M\%S"), ".log"), error = paste0("seatrackRtools_log_ERROR_",
    format(Sys.time(), "\%Y-\%m-\%d_\%H\%M\%S"), ".log"), success =
    paste0("seatrackRtools_log_SUCCESS_", format(Sys.time(), "\%Y-\%m-\%d_\%H\%M\%S"),
    ".log"))
)
}
\arguments{
\item{silent}{A boolean indicating whether to start logs silently.}

\item{log_path}{A string specifying the path to the directory where log files should be stored. If NULL, it defaults to a "seatrackRtools_app/seatrackRtools_logs" directory in the current working directory.}

\item{log_names}{A list of strings specifying the names of the log files for general logs, errors, and successes. Each name should include a timestamp to ensure uniqueness. Defaults to a list with names "seatrackRtools_log_\if{html}{\out{<timestamp>}}.log", "seatrackRtools_log_ERROR_\if{html}{\out{<timestamp>}}.log", and "seatrackRtools_log_SUCCESS_\if{html}{\out{<timestamp>}}.log".}
}
\value{
A list of log file names that were set up for the app.
}
\description{
Set up logs for shiny app
This function initializes logging for the Shiny app, creating separate log files for general logs, errors, and successes. It ensures that the specified log directory exists and starts logging with appropriate settings for each log type.
}
\concept{shiny_app}
\keyword{internal}
