% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/startups.R
\name{add_loggers_from_startup}
\alias{add_loggers_from_startup}
\title{Attempt to add logger from startup sheets}
\usage{
add_loggers_from_startup(master_import, new_metadata)
}
\arguments{
\item{master_import}{Loaded Master startup file.}

\item{new_metadata}{Loaded filled metadata sheet.}
}
\value{
A new version of the master startup data frame, with the logger added if succesful.
}
\description{
This function attempts to add a logger to the master startup data frame from the startup sheets.
Because the data quality of older startup sheets is variable, the function checks for column mismatches and skips these files.
Incorrectly formatted datetime columns can also lead to issues.
}
\examples{
\dontrun{
updated_master_startup <- add_loggers_from_startup_sheets(master_startup)
}
}
\concept{startups}
