% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gls_workflows.R
\name{gls_process_positions}
\alias{gls_process_positions}
\title{Process GLS position data using seatrackRgls}
\usage{
gls_process_positions(species, colony, import_directory, output_directory)
}
\arguments{
\item{species}{Species name for colony information.}

\item{colony}{Colony name for colony information.}

\item{import_directory}{Path to the directory containing GLS files.}

\item{output_directory}{Path to the directory where the processed position data will be saved.}
}
\value{
None. The function saves the processed position data to the specified output directory.
}
\description{
Function to process GLS position data from a specified directory using seatrackRgls.
}
\concept{gls_helper}
