% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gls_workflows.R
\name{gls_process_positions}
\alias{gls_process_positions}
\title{Process GLS position data using seatrackRgls}
\usage{
gls_process_positions()
}
\value{
None. The function saves the processed position data to \verb{Database\\\\Imports_Logger data\\\\Output_GLSpositions}
}
\description{
Function to process calibrated GLS position data using seatrackRgls.
}
\concept{gls_helper}
