% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activity_db_import.R
\name{load_activity}
\alias{load_activity}
\title{Load activity data from file}
\usage{
load_activity(file_info, recording_type)
}
\arguments{
\item{file_info}{A dataframe containing information about the file to be processed, including session_id, filename, individ_id, deployment_date, retrieval_date, full_path, and extension.}

\item{recording_type}{A list containing information about the recording type, including table_name, process_function, extensions, and argname.}
}
\value{
A dataframe containing the clipped activity data for the specified file and recording type, or NULL if there was an error in processing the file or if there is no data within the deployment and retrieval dates.
}
\description{
This function loads activity data from a specified file based on the recording type. It uses the appropriate processing function for the recording type to read the data, clips the data to the deployment and retrieval dates, and returns the clipped activity data as a dataframe. If there is an error in processing the file or if there is no data within the deployment and retrieval dates, it logs a warning and returns NULL.
}
\concept{activity_db_import}
