% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annual_report_tables.R
\name{make_annual_report_tables}
\alias{make_annual_report_tables}
\title{Export tables for annual report}
\usage{
make_annual_report_tables(
  field_plan_path,
  export_dir,
  all_metadata = NULL,
  new_locations = NULL,
  field_year = c(as.numeric(format(Sys.Date(), "\%Y")))
)
}
\arguments{
\item{field_plan_path}{Path to field plan Excel file}

\item{export_dir}{Directory to export tables to}

\item{all_metadata}{Optional data frame of all metadata, if not provided will be loaded from master import}

\item{new_locations}{Optional dataframe of new locations to check against field plan}

\item{field_year}{Year to generate tables for, default is current year}
}
\description{
This function generates tables for the annual report based on the field plan, metadata and database.
}
\concept{annual_report}
