% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_import_prep.R
\name{prepare_session_deployments}
\alias{prepare_session_deployments}
\title{Get session deployments}
\usage{
prepare_session_deployments(session_batch, metadata)
}
\arguments{
\item{session_batch}{A SessionBatch object containing session information from master import startup_shutdown.}

\item{metadata}{A dataframe containing metadata information for the sessions.}
}
\value{
A list containing a dataframe of deployments and a tibble of sessions.
}
\description{
Function to get deployments associated with a set of logger sessions. The function will look for deployments that occurred between the session start time and the session end time (download/shutdown).
If the session is open (no download/shutdown date), the function will use a date far in the future to indicate the session is still open.
If there are ambiguities (multiple deployments for a single session), those sessions will be removed and a warning will be logged.
}
\concept{db_import}
