% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_import.R
\name{push_db_import_collections}
\alias{push_db_import_collections}
\title{Push multiple database import collections to database}
\usage{
push_db_import_collections(db_import_collections)
}
\arguments{
\item{db_import_collections}{A list of DBImportCollection objects.}
}
\description{
Function to push multiple DBImportCollection objects to the database. Each collection is processed in turn using the \code{push_db_import_collection} function.
}
\concept{db_import}
