% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_import.R
\name{push_retrievals}
\alias{push_retrievals}
\title{Push retrievals to database

Function to push retrievals from a master import sheet to the database. It checks the database for existing retrievals and only imports new retrievals to avoid duplicates.
It also checks that the loggers being retrieved have an associated deployment in the database.}
\usage{
push_retrievals(retrievals)
}
\arguments{
\item{retrievals}{A tibble containing retrieval information from a DBImportCollection object.}
}
\description{
Push retrievals to database

Function to push retrievals from a master import sheet to the database. It checks the database for existing retrievals and only imports new retrievals to avoid duplicates.
It also checks that the loggers being retrieved have an associated deployment in the database.
}
\concept{db_import}
