% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_import.R
\name{push_startup}
\alias{push_startup}
\title{Push startups to database}
\usage{
push_startup(startup)
}
\arguments{
\item{startup}{A tibble containing session information from a DBImportCollection object.}
}
\description{
Function to push startups from a master import sheet to the database. It checks the database for existing startups and only imports new startups to avoid duplicates.
}
\concept{db_import}
