% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ColonyRepr_supporting_functions.R
\name{run_ColonyRepr}
\alias{run_ColonyRepr}
\title{Function to extract desired data from the SEATRACK database and fully run the colony representativeness analysis}
\usage{
run_ColonyRepr(
  species,
  colony,
  Month_of_interest = 12,
  KDE_contours_of_interest = seq(from = 25, to = 95),
  smoothing_parameter = NULL,
  N_iterations = 20
)
}
\arguments{
\item{species}{Specify species of interest (numeric)}

\item{colony}{Specify colony of interest (numeric)}

\item{Month_of_interest}{Specify month of interest (numeric)}

\item{KDE_contours_of_interest}{Specify KDE contours of interest (can be a vector)}

\item{smoothing_parameter}{Specify a smoothing parameter to be used in the UD calculation}

\item{N_iterations}{Specify the number of iterations you'd like to perform when combining surface areas (the mean will be used in fitting the MM-curve)}
}
\value{
A list with all components of the colony representativeness analysis
}
\description{
Function to extract desired data from the SEATRACK database and fully run the colony representativeness analysis
}
\concept{ColonyRepr}
