% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/startups.R
\name{search_startup_files}
\alias{search_startup_files}
\title{Search for logger instances in startup files}
\usage{
search_startup_files(target_logger_ids, existing_id_date, master_startup)
}
\arguments{
\item{target_logger_ids}{A character vector of logger IDs to search for in the startup files.}

\item{existing_id_date}{A character vector of existing logger ID-date combinations to avoid duplicates.}
}
\value{
A tibble containing the logger instances found in the startup files that are not already present in the existing ID-date combinations.
}
\description{
This function searches for logger instances in startup files based on the provided logger IDs and existing ID-date combinations.
}
\concept{startups}
