% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{setup_app_logs}
\alias{setup_app_logs}
\title{Set up logs for shiny app}
\usage{
setup_app_logs(
  silent = FALSE,
  log_path = NULL,
  log_names = list(global = paste0("seatrackRtools_log_", format(Sys.time(),
    "\%Y-\%m-\%d_\%H\%M\%S"), ".txt"), error = paste0("seatrackRtools_log_ERROR_",
    format(Sys.time(), "\%Y-\%m-\%d_\%H\%M\%S"), ".txt"))
)
}
\description{
Set up logs for shiny app
}
\concept{shiny_app}
